+++ src/redturtle/prenotazioni/interfaces.py
@@ -2,6 +2,7 @@
 """Module where all interfaces, events and exceptions live."""
 
 from zope.interface import Interface
+from zope.interface.interfaces import IObjectEvent
 from zope.publisher.interfaces.browser import IDefaultBrowserLayer
 
 
@@ -23,3 +24,7 @@ class IPrenotazioneEmailMessage(Interface):
 
 class ISerializeToRetroCompatibleJson(Interface):
     """Interface used to cereate the TEMPORARY retrocomattible serializers"""
+
+
+class IBookingReminderEvent(IObjectEvent):
+    """Booking reminder time arrived event"""